% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{bam}
\alias{bam}
\alias{bam,CRCViewList-method}
\alias{bam,CRCView-method}
\title{Get bam}
\usage{
bam(object)

\S4method{bam}{CRCViewList}(object)

\S4method{bam}{CRCView}(object)
}
\arguments{
\item{object}{Object holding bam paths.}
}
\description{
Get bam
}
