% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R
\docType{class}
\name{Clique-class}
\alias{Clique-class}
\alias{hash}
\alias{members}
\title{A class for holding groups of TranscriptionFactor objects.}
\usage{
\S4method{hash}{Clique}(clique)

\S4method{members}{Clique}(clique)
}
\description{
Contains at least one TranscriptionFactor object, as well as
a list of names of contained TF instances.
}
