% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R
\docType{class}
\name{CliqueList-class}
\alias{CliqueList-class}
\alias{unique_cliques}
\alias{extract_cliques}
\alias{name}
\alias{name}
\alias{hashes}
\title{A class for holding groups of Clique objects.}
\usage{
\S4method{unique_cliques}{CliqueList}(object)

\S4method{extract_cliques}{CliqueList}(object)

\S4method{name}{CliqueList}(x)

\S4method{name}{CliqueList,character}(x) <- value

\S4method{hashes}{CliqueList}(cliquelist)
}
\description{
A class for holding groups of Clique objects.
}
