% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{combine}
\alias{combine}
\alias{combine,list-method}
\alias{combine,TranscriptionFactor-method}
\title{Combine TranscriptionFactor instances.}
\usage{
combine(x, ..., name = name(x))

\S4method{combine}{list}(x, ..., name = NULL)

\S4method{combine}{TranscriptionFactor}(x, ..., name = NULL)
}
\arguments{
\item{x}{A TranscriptionFactor object or list of TranscriptionFactor objects.}

\item{...}{TranscriptionFactor objects.}

\item{name}{A new name for the combined TF. Optional. Defaults to name of first TF.}
}
\description{
Combine TranscriptionFactor instances.
}
