% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstreamHelpers.R
\name{combine_similar_vectors}
\alias{combine_similar_vectors}
\title{Combine similar vecs}
\usage{
combine_similar_vectors(named_list_of_vectors, max_diff_to_combine = 1,
  verbose = F)
}
\arguments{
\item{named_list_of_vectors}{list of vecs}

\item{max_diff_to_combine}{recursively combine until all cliques are have more diffs than this arg.}

\item{verbose}{print progress for debugging}
}
\description{
Not really intended to be called by user.
}
\details{
Uses doparallel backend, register this with
'registerDoParallel(12)'.
}
