% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CRCExperiment-class}
\alias{CRCExperiment-class}
\alias{CRCExperiment}
\title{A class for containing the full representation of a cohort's coltron
results.}
\usage{
CRCExperiment(rse, crclist, cohort = NULL)
}
\arguments{
\item{rse}{A RangedSummarizedExperiment object.}

\item{crclist}{A CRCList object.}

\item{cohort}{Optional name for experimental group.}
}
\description{
A class for containing the full representation of a cohort's coltron
results.
}
