% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CRCView-class}
\alias{CRCView-class}
\alias{CRCView}
\title{A class for holding a representation of a single sample's predicted
network.}
\usage{
CRCView(quantsites, cliques, prior_tfbs = NULL, sample = name(cliques),
  bampath = NULL)
}
\arguments{
\item{quantsites}{GRanges object.}

\item{cliques}{CliqueList object.}

\item{prior_tfbs}{Optional GrangesList of predicted TFBS.}

\item{sample}{Sample name.}

\item{bampath}{Path to bam for quantification.}
}
\description{
Contains a GRanges object holding accessible regions for downstream consideration.
Additionally may contain a name, a CliqueList, all predicted TFBS, and a path
to a bam used for quantification.
}
