% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CRCViewList-class}
\alias{CRCViewList-class}
\alias{CRCViewList}
\title{A class for holding a list of CRCViews.}
\usage{
CRCViewList(crcview, ..., name = NULL)
}
\arguments{
\item{crcview}{A CRCView object.}

\item{...}{More CRCView Objects.}

\item{name}{A name for the cohort.}
}
\description{
A class for holding a list of CRCViews.
}
