% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{extract_cliques}
\alias{extract_cliques}
\alias{unique_cliques}
\alias{unique_cliques,CliqueList-method}
\alias{extract_cliques,CliqueList-method}
\alias{extract_cliques,CRCView-method}
\alias{extract_cliques,CRCViewList-method}
\alias{extract_cliques,CRCExperiment-method}
\title{Clique extractors.}
\usage{
extract_cliques(object)

unique_cliques(object)

\S4method{unique_cliques}{CliqueList}(object)

\S4method{extract_cliques}{CliqueList}(object)

\S4method{extract_cliques}{CRCView}(object)

\S4method{extract_cliques}{CRCViewList}(object)

\S4method{extract_cliques}{CRCExperiment}(object)
}
\arguments{
\item{object}{Object holding cliques.}
}
\description{
Clique extractors.
}
