% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstreamHelpers.R
\name{get_diff_cliques}
\alias{get_diff_cliques}
\title{Run differential analysis at the clique level.}
\usage{
get_diff_cliques(object, ix, contrast = NULL, parametric = TRUE,
  score_type = c("dev", "z"))
}
\arguments{
\item{object}{A CRCExperiment object.}

\item{ix}{A logical matrix holding clique index information.}

\item{contrast}{Optionally provide vector containing 2 levels of metadata CONDITION.}

\item{parametric}{Passed to \link[chromVAR]{differentialDeviations}.}

\item{score_type}{Use chromVAR's deviations or deviation z-scores as metric for results table.}
}
\value{
S3 object of class CRCResult, a list holding means+pvalues, sample scores, and a chromVARdeviations object.
}
\description{
Supply a CRCExperiment, and index matrix.
}
\details{
See \link[chromVAR]{computeDeviations}.
See \link[chromVAR]{differentialDeviations}.
}
