% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{hash}
\alias{hash}
\alias{hashes}
\alias{hash,Clique-method}
\alias{hashes,CliqueList-method}
\title{Get unique hash.}
\usage{
hash(clique)

hashes(cliquelist)

\S4method{hash}{Clique}(clique)

\S4method{hashes}{CliqueList}(cliquelist)
}
\arguments{
\item{clique}{A Clique.}

\item{cliquelist}{A CliqueList.}
}
\description{
Get unique hash.
}
