% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectHelpers.R
\name{list_to_cliquelist}
\alias{list_to_cliquelist}
\title{Convert list of lists containing TF names to a CliqueList.}
\usage{
list_to_cliquelist(cliques)
}
\arguments{
\item{cliques}{A list of character vectors.}
}
\description{
Convert list of lists containing TF names to a CliqueList.
}
