% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{name}
\alias{name}
\alias{name<-}
\alias{members}
\alias{name,TranscriptionFactor-method}
\alias{name,CliqueList-method}
\alias{name,CRCView-method}
\alias{name<-,TranscriptionFactor,character-method}
\alias{name<-,CliqueList,character-method}
\alias{members,Clique-method}
\title{Name getters.}
\usage{
name(x)

name(x) <- value

members(clique)

\S4method{name}{TranscriptionFactor}(x)

\S4method{name}{CliqueList}(x)

\S4method{name}{CRCView}(x)

\S4method{name}{TranscriptionFactor,character}(x) <- value

\S4method{name}{CliqueList,character}(x) <- value

\S4method{members}{Clique}(clique)
}
\arguments{
\item{x}{A name-able object.}

\item{value}{New name.}

\item{clique}{A clique object.}
}
\description{
Name getters.
}
