% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstreamHelpers.R
\name{plot_cliques}
\alias{plot_cliques}
\title{Plotting helper for CRCResult object.}
\usage{
plot_cliques(object, plot = c("volcano", "tsne", "clustergram"),
  contrast = NULL, use.adjusted.p = F)
}
\arguments{
\item{object}{An object of class CRCResult.}

\item{plot}{A string specifying the type of plot to produce. Only 'volcano' implemented for now.}

\item{contrast}{A vector containing the 2 levels of CONDITION to compare.}

\item{use.adjusted.p}{Logical specifying whether to use adjusted or non-adjusted p-values for plotting.}
}
\value{
ggplot2 object.
}
\description{
Plotting helper for CRCResult object.
}
