% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{pwms}
\alias{pwms}
\alias{pwms,TranscriptionFactor-method}
\title{Get pwms.}
\usage{
pwms(x)

\S4method{pwms}{TranscriptionFactor}(x)
}
\arguments{
\item{x}{A TranscriptionFactor object.}
}
\description{
Get pwms.
}
