% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{quantsites}
\alias{quantsites}
\alias{quantsites,CRCView-method}
\alias{quantsites,CRCViewList-method}
\title{Get quantsites.}
\usage{
quantsites(object)

\S4method{quantsites}{CRCView}(object)

\S4method{quantsites}{CRCViewList}(object)
}
\arguments{
\item{object}{Object holding quantsites.}
}
\description{
Get quantsites.
}
