% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importColtron.R
\name{read_coltron_subpeaks}
\alias{read_coltron_subpeaks}
\title{Import subpeak regions from a coltron run.}
\usage{
read_coltron_subpeaks(path)
}
\arguments{
\item{path}{A path to a coltron output directory.}
}
\value{
GRanges object.
}
\description{
Usually not run by user.
}
