% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{remove_subset_cliques}
\alias{remove_subset_cliques}
\alias{remove_subset_cliques,CliqueList-method}
\alias{remove_subset_cliques,CRCExperiment-method}
\title{Remove cliques that are subsets of other cliques}
\usage{
remove_subset_cliques(object)

\S4method{remove_subset_cliques}{CliqueList}(object)

\S4method{remove_subset_cliques}{CRCExperiment}(object)
}
\description{
Remove cliques that are subsets of other cliques
}
