% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstreamHelpers.R
\name{remove_subset_vectors}
\alias{remove_subset_vectors}
\title{Return vectors that are not subsets of other vectors in obj}
\usage{
remove_subset_vectors(named_list_of_vectors)
}
\description{
Return vectors that are not subsets of other vectors in obj
}
