% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,CRCView-method}
\alias{show,CRCExperiment-method}
\title{Print object representation.}
\usage{
show(object)

\S4method{show}{CRCView}(object)

\S4method{show}{CRCExperiment}(object)
}
\arguments{
\item{object}{Object with printable representation.}
}
\description{
Print object representation.
}
