% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{tfbs}
\alias{tfbs}
\alias{tfbs,CRCView-method}
\alias{tfbs,CRCViewList-method}
\alias{tfbs,CRCExperiment-method}
\title{Get tfbs.}
\usage{
tfbs(object)

\S4method{tfbs}{CRCView}(object)

\S4method{tfbs}{CRCViewList}(object)

\S4method{tfbs}{CRCExperiment}(object)
}
\arguments{
\item{object}{Object holding tfbs as a GRanges object.}
}
\description{
Get tfbs.
}
