% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectHelpers.R
\name{tfbs_by_clique}
\alias{tfbs_by_clique}
\title{Group TFBS  by clique.}
\usage{
tfbs_by_clique(object)
}
\arguments{
\item{object}{A CRCExperiment, CRCViewList, or CRCView object.}
}
\description{
Group TFBS  by clique.
}
