% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_crc}
\alias{toy_crc}
\title{A mini CRCExperiment for vignette building and tests.}
\format{CRCExperiment}
\usage{
toy_crc
}
\description{
Derived from 4x ENCODE datasets processed with coltron.
}
\keyword{datasets}
