% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TranscriptionFactor-class}
\alias{TranscriptionFactor-class}
\alias{TranscriptionFactor}
\title{A class for holding Transcription Factors.}
\usage{
TranscriptionFactor(name, pwms = NULL)
}
\arguments{
\item{name}{A name for the TranscriptionFactor.}

\item{pwms}{Optionally a PWMatrixList object.}
}
\description{
Holds a name and a PWMatrixList, which can be empty.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name for the TF}

\item{\code{pwms}}{PWMatrixList. Defined in TFBSTools.}
}}

