% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Clique-class}
\alias{Clique-class}
\alias{Clique}
\title{A class for holding groups of TranscriptionFactor objects.}
\usage{
Clique(tf, ..., name = NULL)
}
\arguments{
\item{tf}{TranscriptionFactor object.}

\item{...}{more TranscriptionFactor objects.}

\item{name}{A name for the clique.}
}
\description{
Contains at least one TranscriptionFactor object, as well as
a list of names of contained TF instances.
}
