% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique-signal.R
\docType{package}
\name{clique.signal}
\alias{clique.signal}
\alias{clique.signal-package}
\title{clique.signal: A package for working with coltron results.}
\description{
The clique.signal package provided functionality for importing results
generated by coltron.
}
\details{
Coltron was developed by Alex Federation in the Bradner Lab at DFCI.
}
\seealso{
\url{http://www.alexfederation.com/code/},
\url{https://pypi.org/project/coltron/}
}
