% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CliqueList-class}
\alias{CliqueList-class}
\alias{CliqueList}
\title{A class for holding groups of Clique objects.}
\usage{
CliqueList(clique, ..., name = NULL)
}
\arguments{
\item{clique}{Clique object.}

\item{...}{more Clique objects.}

\item{name}{Sample name.}
}
\description{
A class for holding groups of Clique objects.
}
