% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{combine_similar_cliques}
\alias{combine_similar_cliques}
\alias{combine_similar_cliques,CliqueList-method}
\title{Combine cliques that are similar}
\usage{
combine_similar_cliques(object, ...)

\S4method{combine_similar_cliques}{CliqueList}(object,
  combine_when_at_least_as_similar = 1)
}
\arguments{
\item{object}{CliqueList object.}

\item{...}{Other args. See other methods.}

\item{combine_when_at_least_as_similar}{Return cliquelist with <N differences combined.}
}
\description{
Combine cliques that are similar
}
