% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R
\docType{class}
\name{CRCExperiment-class}
\alias{CRCExperiment-class}
\alias{CRCExperiment}
\alias{tfbs}
\alias{extract_cliques}
\alias{show}
\title{A class for containing the full representation of a cohort's coltron
results.}
\usage{
CRCExperiment(rse, crclist, cohort = NULL)

\S4method{tfbs}{CRCExperiment}(object)

\S4method{extract_cliques}{CRCExperiment}(object)

\S4method{show}{CRCExperiment}(object)
}
\description{
A class for containing the full representation of a cohort's coltron
results.
}
