% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R
\docType{class}
\name{CRCView-class}
\alias{CRCView-class}
\alias{CRCView}
\alias{bam}
\alias{quantsites}
\alias{tfbs}
\alias{extract_cliques}
\alias{name}
\alias{show}
\title{A class for holding a representation of a single sample's predicted
network.}
\usage{
CRCView(quantsites, cliques, prior_tfbs = NULL, sample = name(cliques),
  bampath = NULL)

\S4method{bam}{CRCView}(object)

\S4method{quantsites}{CRCView}(object)

\S4method{tfbs}{CRCView}(object)

\S4method{extract_cliques}{CRCView}(object)

\S4method{name}{CRCView}(x)

\S4method{show}{CRCView}(object)
}
\description{
Contains a GRanges object holding accessible regions for downstream consideration.
Additionally may contain a name, a CliqueList, all predicted TFBS, and a path
to a bam used for quantification.
}
