% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R
\docType{class}
\name{CRCViewList-class}
\alias{CRCViewList-class}
\alias{CRCViewList}
\alias{bam}
\alias{quantsites}
\alias{tfbs}
\alias{extract_cliques}
\title{A class for holding a list of CRCViews.}
\usage{
CRCViewList(crcview, ..., name = NULL)

\S4method{bam}{CRCViewList}(object)

\S4method{quantsites}{CRCViewList}(object)

\S4method{tfbs}{CRCViewList}(object)

\S4method{extract_cliques}{CRCViewList}(object)
}
\description{
A class for holding a list of CRCViews.
}
