% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importColtron.R
\name{create_coltron_experiment}
\alias{create_coltron_experiment}
\title{create_coltron_experiment}
\usage{
create_coltron_experiment(metadata, quantsites = "SUBPEAKS")
}
\description{
supply a data.frame containing at least:
SAMPLE, CONDITION, COLTRONDIR, BAM
}
\details{
QUANTSITES is an optional field for specifying atac regions to
search for motifs in and quantify signal. Defaults to coltron's subpeaks
bed file otherwise.
}
