% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstreamHelpers.R
\name{index_by_clique}
\alias{index_by_clique}
\title{Index a CRCExperiment ranges by clique motif presence.}
\usage{
index_by_clique(object)
}
\arguments{
\item{object}{A CRCExperiment object.}
}
\value{
A sparse Matrix holding clique motif presence/absence.
}
\description{
Index a CRCExperiment ranges by clique motif presence.
}
