% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstreamHelpers.R
\name{remove_subset_vectors}
\alias{remove_subset_vectors}
\title{Remove vecs that are perfect subsets of other vecs}
\usage{
remove_subset_vectors(named_list_of_vectors)
}
\arguments{
\item{named_list_of_vectors}{list of vecs}
}
\description{
Not really intended to be called by user.
}
