% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R
\docType{class}
\name{TranscriptionFactor-class}
\alias{TranscriptionFactor-class}
\alias{name}
\alias{name}
\alias{pwms}
\alias{combine}
\alias{combine}
\title{A class for holding Transcription Factors.}
\usage{
\S4method{name}{TranscriptionFactor}(x)

\S4method{name}{TranscriptionFactor,character}(x) <- value

\S4method{pwms}{TranscriptionFactor}(x)

\S4method{combine}{list}(x, ..., name = NULL)

\S4method{combine}{TranscriptionFactor}(x, ..., name = NULL)
}
\description{
Holds a name and a PWMatrixList, which can be empty.
}
